New Parallax Background
I added textures for Level 1 Background Layer 0/1 (sky and mountains)
and Layer 2 (hills) into the assets folder. Also, I created a texture
that combines Level 1 Background Layers 0, 1, and 2 for temporary use. I
changed the parallax script to alter the delta x and delta y for the new
texture. The next commit will seperate the movement of Layers 1 and 2.